New tank velocity inheritance system
With this new system, a player will inherit it's tank's velocity, as long as they are within X units of the cell that they dismounted from. Once you get over X distance away from the tank, you are now unparented from the tank, and you have your own independent velocity. This is visualized with a circle that goes from blue to red dependent on distance, which can be seen in the scene view when dismounting a tank. Once you stand on top of a tank again, or go inside of a tank, you are back to being influenced by that tank's velocity until you fully dismount again

as long as you are not in the tank but within your dismount point radius, you have "soft dismounted", and once you leave that circle, you have "full dismounted"