Hunt state overhaul & substate system changes
When hunt state is entered, if the enemy tank has a weapon already active, it will look at that weapon, and then see if there are any higher priority weapons that are available to switch to. If there are, it'll switch it's token to that weapon, and shoot at the player. So, in the instance of that clip, switching from mortar to machine gun to hose the player. When the hunt state is exited, it will return the switched token back to where it was (would go back to mortar). If there is nothing better to switch to, but we have a weapon active, it'll just use that weapon.

If there is an open token at all, and it doesn't have to switch, it'll just man the highest priority gun possible.

If there are no open tokens, and no guns being manned, it'll switch a token from boiler to the priority gun.

In engage state, when tokens are redistributed, it will keep the interactable that's hunting the player active. It will also do this when changing parent states. So, when it re-evaluates it's tokens after the switch, if hunt is active, it will re-distribute with one less token in mind.

Any interactables being overridden by a substate now have a magenta gizmo box around them, to differentiate them for debug reasons

Changed substates so that the substate isn't exited and then entered again if the substate is currently active while switching parent states, as long as the state it's trying to switch to also has the active substate as a child