Made new aiming algorithm, MUCH better now
When our weapon's projected hitpoint goes past it's target, we don't want to use the exact projected hitpoint anymore because it can't be reliably used at that point.

Before, in this scenario, I just set the projected hitpoint a few meters in front of the barrel instead, which honestly didnt work well at all. Was slightly more accurate than using the unreliable hitpoint, but caused a lot of issues, and made it so aim was always shakey in that scenario.

Now, in this scenario, we instead get the point where our trajectory intersects with the Y axis of our target point, and we use THAT as our projected hit point. This is the best and most accurate way to ensure the weapons are always hitting their target.

Also improved aim to be much less shakey / jittery at all times