Enemies cant man broken interactables, or interactables manned by a player, and new surrender con
If a player mans an interactable the AI is currently using, it'll stop using it until the player leaves the interactable. If an interactable the AI is using breaks, it will stop using that until the interactable is fixed. Tank will now surrender if it has no guns, or if it only has guns left that are being controlled by players