Added new hatch system for multiple hatches (not quite done, but stable), added runtime polaroid system
random hatches dont show up in the game scene yet, but they work in the build scene and populate with no errors

So this is an overhaul to the hatch system that allows for multiple different random hatch rooms to spawn in the party box. This required creating a new polaroid system for creating sprites of objects at runtime, as well as new systems for spawning random hatches in valid cells, had to add a lot of code to room.cs and the room building scripts to get this to work. this was a lot of work honestly.